add offset stage scanner using the new baseclasses, and remove raster_scan_camera.py
